+++ src/TreeTools-init.c
@@ -9,10 +9,12 @@ extern SEXP _TreeTools_as_newick(SEXP);
 extern SEXP _TreeTools_consensus_tree(SEXP, SEXP);
 extern SEXP _TreeTools_cpp_edge_to_splits(SEXP, SEXP);
 extern SEXP _TreeTools_drop_tip(SEXP, SEXP);
+extern SEXP _TreeTools_edge_to_mixed_base(SEXP, SEXP, SEXP);
 extern SEXP _TreeTools_edge_to_num(SEXP, SEXP, SEXP);
 extern SEXP _TreeTools_edge_to_rooted_shape(SEXP, SEXP, SEXP);
 extern SEXP _TreeTools_keep_tip(SEXP, SEXP);
 extern SEXP _TreeTools_minimum_spanning_tree(SEXP);
+extern SEXP _TreeTools_mixed_base_to_parent(SEXP, SEXP);
 extern SEXP _TreeTools_num_to_parent(SEXP, SEXP);
 extern SEXP _TreeTools_postorder_edges(SEXP, SEXP);
 extern SEXP _TreeTools_preorder_edges_and_nodes(SEXP, SEXP);
@@ -35,10 +37,12 @@ static const R_CallMethodDef callMethods[] = {
   {"_TreeTools_consensus_tree", (DL_FUNC) &_TreeTools_consensus_tree, 2},
   {"_TreeTools_cpp_edge_to_splits", (DL_FUNC) &_TreeTools_cpp_edge_to_splits, 2},
   {"_TreeTools_drop_tip", (DL_FUNC) &_TreeTools_drop_tip, 2},
+  {"_TreeTools_edge_to_mixed_base", (DL_FUNC) &_TreeTools_edge_to_mixed_base, 3},
   {"_TreeTools_edge_to_num", (DL_FUNC) &_TreeTools_edge_to_num, 3},
   {"_TreeTools_edge_to_rooted_shape", (DL_FUNC) &_TreeTools_edge_to_rooted_shape, 3},
   {"_TreeTools_keep_tip", (DL_FUNC) &_TreeTools_keep_tip, 2},
   {"_TreeTools_minimum_spanning_tree", (DL_FUNC) &_TreeTools_minimum_spanning_tree, 1},
+  {"_TreeTools_mixed_base_to_parent", (DL_FUNC) &_TreeTools_mixed_base_to_parent, 2},
   {"_TreeTools_num_to_parent", (DL_FUNC) &_TreeTools_num_to_parent, 2},
   {"_TreeTools_postorder_edges", (DL_FUNC) &_TreeTools_postorder_edges, 2},
   {"_TreeTools_preorder_edges_and_nodes", (DL_FUNC) &_TreeTools_preorder_edges_and_nodes, 2},